feat: 15.1

+++ 04-Acme_Santiago/Workspace/Acme-Santiago/src/main/java/services/ChirpService.java
@@ -1,12 +1,15 @@
 package services;
 
+import java.util.ArrayList;
 import java.util.Collection;
-
-import javax.transaction.Transactional;
+import java.util.Date;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.Validator;
 
 import repositories.ChirpRepository;
 import domain.Chirp;
@@ -16,39 +19,55 @@ import domain.User;
 @Transactional
 public class ChirpService {
 
-	// Managed repository
+	// Managed repository -----------------------------------------------------
+
 	@Autowired
 	private ChirpRepository chirpRepository;
 
-	// Suporting services
+	@Autowired
+	private AdministratorService administratorService;
+	// Supporting services ----------------------------------------------------
+	
 	@Autowired
 	private UserService userService;
+	
+	@Autowired
+	private ConfigurationService configurationService;
+	
+	@Autowired
+	private Validator validator;
 
-	// Constructors
+	// Constructor ------------------------------------------------------------
 
 	public ChirpService() {
 		super();
 	}
 
-	// Simple CRUD methods
+	// Simple CRUD methods ----------------------------------------------------
 
 	public Chirp create() {
-		Chirp res;
-		res = new Chirp();
-
-		User user = userService.findByPrincipal();
-		res.setUser(user);
-
-		return res;
+		this.userService.checkAuthority();
+		User user;
+		Chirp result;
+		Date moment;
+		
+		result = new Chirp();
+		moment = new Date(System.currentTimeMillis() - 1000);
+		user = userService.findByPrincipal();
+		result.setMoment(moment);
+		result.setUser(user);
+		
+		return result;
 	}
 
 	public Collection<Chirp> findAll() {
 		Collection<Chirp> res;
 		res = this.chirpRepository.findAll();
+		Assert.notNull(res);
 		return res;
 	}
 
-	public Chirp findOne(final int id) {
+	public Chirp findOne(int id) {
 		Assert.isTrue(id != 0);
 		Chirp res;
 		res = this.chirpRepository.findOne(id);
@@ -56,24 +75,79 @@ public class ChirpService {
 		return res;
 	}
 
-	public Chirp save(final Chirp chirp) {
+	public Chirp save(Chirp chirp) {
 		Assert.notNull(chirp);
 		Chirp res;
-
 		res = this.chirpRepository.save(chirp);
 		return res;
 	}
 
 	public void delete(Chirp chirp) {
+		this.administratorService.checkAuthority();
 		Assert.notNull(chirp);
 		Assert.isTrue(chirp.getId() != 0);
 		Assert.isTrue(this.chirpRepository.exists(chirp.getId()));
 		this.chirpRepository.delete(chirp);
 	}
 
-	// Other business methods
+	// Other business method --------------------------------------------------
+	
+	public Collection<Chirp> findChirpByUser(int id) {
+		Collection<Chirp> res;
+		res = this.chirpRepository.findChirpByUser(id);
+		return res;
+	}
+	
+	public void checkTabooWords() {
+		Collection<String> tabooWords = new ArrayList<String>();
+		tabooWords = configurationService.findTabooWords();
+
+		Collection<Chirp> chirps = new ArrayList<Chirp>();
+		chirps = this.findAll();
+
+		for (String s : tabooWords) {
+			for (Chirp c : chirps) {
+				if (c.getTitle().toLowerCase().contains(s.toLowerCase()) || c.getText().toLowerCase().contains(s.toLowerCase())) {
+					c.setTaboo(true);
+				}
+			}
+		}
+	}
+	
+	public Collection<Chirp> findChirpTaboo(){
+		Collection<Chirp> res = new ArrayList<Chirp>();
+		res.addAll(chirpRepository.findChirpTaboo());
+		return res;
+	}
 
 	public void flush() {
 		this.chirpRepository.flush();
 	}
+	
+	public Chirp reconstruct(final Chirp chirp, final BindingResult binding) {
+		Chirp res;
+		Chirp chirpFinal;
+		if (chirp.getId() == 0) {
+			User userPrincipal;
+
+			userPrincipal = this.userService.findByPrincipal();
+			chirp.setUser(userPrincipal);
+			
+			res = chirp;
+		} else {
+			chirpFinal = this.chirpRepository.findOne(chirp.getId());
+			
+			chirp.setId(chirpFinal.getId());
+			chirp.setVersion(chirpFinal.getVersion());
+			chirp.setMoment(chirpFinal.getMoment());
+			chirp.setUser(chirpFinal.getUser());
+			chirp.setText(chirpFinal.getText());
+			chirp.setTitle(chirpFinal.getTitle());
+
+			res = chirp;
+		}
+		this.validator.validate(res, binding);
+		return res;
+	}
+
 }
